+++ src/chunk.c
@@ -84,6 +84,7 @@ const char *opcode_nombre(OpCode op) {
         case OP_ITER_SIGUIENTE:  return "OP_ITER_SIGUIENTE";
         case OP_FORMATO_F:       return "OP_FORMATO_F";
         case OP_ASEGURAR_CADENA: return "OP_ASEGURAR_CADENA";
+        case OP_LONGITUD:        return "OP_LONGITUD";
         case OP_RETORNAR:        return "OP_RETORNAR";
     }
     return NULL;